src/policies-search/pyproject.toml
[project]
name = "policies-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.14",
    "azure-identity>=1.23.0",
    "azure-search-documents>=11.5.3",
    "azure-storage-blob>=12.19.0",
    "python-dotenv>=1.1.1",
]
